Split CLI runtime orchestration from entry point
