+++ packages/agent-cli/agent-cli.cabal
@@ -56,6 +56,8 @@ library
                     , Agent.CLI.PrivateFileLock
                     , Agent.CLI.Runtime
                     , Agent.CLI.Runtime.Internal
+                    , Agent.CLI.Runtime.Orchestration
+                    , Agent.CLI.Runtime.Orchestration.Types
                     , Agent.CLI.Runtime.HistorySource
                     , Agent.CLI.Runtime.Persistence
                     , Agent.CLI.Runtime.Recap
